+++ src/main/java/net/silentchaos512/treasurebags/lib/Const.java
@@ -5,6 +5,7 @@ import net.silentchaos512.treasurebags.TreasureBags;
 
 public class Const {
     public static final class Bags {
+        public static final ResourceLocation DEFAULT = get("default");
         public static final ResourceLocation DUNGEON = get("dungeon");
         public static final ResourceLocation ENDER = get("ender");
         public static final ResourceLocation FOOD = get("food");
@@ -14,6 +15,7 @@ public class Const {
         public static final ResourceLocation PLAYER = get("player");
         public static final ResourceLocation SPAWN = get("spawn");
         public static final ResourceLocation STICKS_AND_STONES = get("sticks_and_stones");
+        public static final ResourceLocation TEST = get("test");
 
         private Bags() {}
     }
@@ -21,6 +23,7 @@ public class Const {
     public static final class LootTables {
         public static final ResourceLocation STARTING_INVENTORY = get("starting_inventory");
 
+        public static final ResourceLocation BAGS_DEFAULT = get("bags/default");
         public static final ResourceLocation BAGS_DUNGEON = get("bags/dungeon");
         public static final ResourceLocation BAGS_ENDER = get("bags/ender");
         public static final ResourceLocation BAGS_FOOD = get("bags/food");
@@ -30,6 +33,7 @@ public class Const {
         public static final ResourceLocation BAGS_PLAYER = get("bags/player");
         public static final ResourceLocation BAGS_SPAWN = get("bags/spawn");
         public static final ResourceLocation BAGS_STICKS_AND_STONES = get("bags/sticks_and_stones");
+        public static final ResourceLocation BAGS_TEST = get("bags/test");
 
         private LootTables() {}
     }