+++ CHANGELOG.md
@@ -4,7 +4,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Bag groups and a config option to disable bag types by group [#6]
 ### Changed
+- Improved logging of errored bag types
 - Switched config to use Forge config system
 
 ## [1.3.1] - 2020-12-17